Implement REST controllers and services for Next Step identity configuration

+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/entity/OtpPolicyDetail.java
@@ -35,8 +35,8 @@ public class OtpPolicyDetail {
     private String otpPolicyName;
     private String description;
     @NotNull
-    private Long length;
-    private Long attemptLimit;
+    private Integer length;
+    private Integer attemptLimit;
     private Long expirationTime;
     private String genAlgorithm;
     @NotNull
