+++ powerauth-nextstep-model/src/main/java/io/getlime/security/powerauth/lib/nextstep/model/request/CreateOtpPolicyRequest.java
@@ -31,8 +31,8 @@ public class CreateOtpPolicyRequest {
     private String otpPolicyName;
     private String description;
     @NotNull
-    private Long length;
-    private Long attemptLimit;
+    private Integer length;
+    private Integer attemptLimit;
     private Long expirationTime;
     private String genAlgorithm;
